graphing/requirements.txt
locustio==0.7.5
influxdb==3.0.0
pika==0.10.0
